Adding camera follow and proper shake, + reorganizing scripts - 22/12
