+++ Assets/Resources/Cinemachine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: abda243a1dd0d0e43b8f8aa85bd86299
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
