+++ Assets/Resources/Cinemachine/RewindCameraShakeNoise.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6061ae6e16ce41d4c9de6633b454ba82
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
