+++ Assets/Resources/Scripts/Camera/CameraRewindShake.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b162e3cc33167ae49b5819a4698137a9
+guid: c86e6516aaa9e66428df706a51d4ca9c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
