--- Assets/Resources/Scripts/Ladder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a8886907c53b6d4bb20563a76f769c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
