--- Assets/Resources/Scripts/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: abef1bb4f6569784b870ed96603f3cea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
