--- Assets/Resources/Scripts/SimpleDoor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 24db5288321a5f348abd2e43db6e8cec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
