--- Assets/Resources/Scripts/Switch.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d6ccd089981d4774da384c26ae99801f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
