+++ Assets/Resources/Scripts/TimeClones/Clone.cs.meta
